ndmp: avoid isilon 9.1.0.0 ndmp problem
The isilon 9.1.0.0 release seems to have added a sanity check in ndmp
that checks that the tape offsets are on 512B boundaries.

The following error is issued when a restore is triggered to a isilon
9.1.0.0 system:

  'Invalid nlist.tape_offset -1 at index 1 - tape offset not aligned at 512B boundary'

However, -1 is an allowed value (NDMP_INVALID_U_QUAD) and was accepted
on every NDMP implementation we tested against including older isilon
versions.

To fix the problem, the isilon needs to check for NDMP_INVALID_U_QUAD
before it checks for the 512B boundary to fix the problem.

Otherwise we could enable this workaround depending on a new
configuration option.